add Duration > preparation for the feature > adopted a feature centric file organisation
