+++ src/common/timeFormats.js
@@ -1,0 +1,4 @@
+const timeFormats = {
+  iso: "YYYY-MM-DDTHH:mm:00+02:00",
+  year: "YYYY-MM-DD",
+};
