+++ src/features/add_event/inputs/EventLabel.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function EventLabel({ children }) {
+  return <label className="block w-full leading-relaxed tracking-wider text-blue-700 ">{children}</label>;
+}
+
+export default EventLabel;
