+++ src/features/add_event/inputs/index.js
@@ -1,0 +1,7 @@
+// COMPONENTS
+import EventLabel from "./EventLabel";
+import AllDayEventInput from "./AllDayEventInput";
+import TimeInput from "./TimeInput";
+import Toggle from "./Toggle";
+
+export { AllDayEventInput, TimeInput, Toggle, EventLabel };
